fix up haha
forgot these

+++ source/shared-code/utilities/string.cpp
@@ -205,8 +205,8 @@ namespace utilities::string
 		size_t offset = to_lower(text).find(to_lower(input));
 		if (offset == std::string::npos) return 0.00; // mismatch
 
-		int len_variance = text.length() - input.length();
-		int match_percent = 100 - (1 + len_variance + offset);
+		size_t len_variance = text.length() - input.length();
+		size_t match_percent = 100 - (1 + len_variance + offset);
 
 		return ((double)match_percent / 100);
 	}
